refactor: update vip number
